+++ src/pages/Auth.jsx
@@ -1,15 +1,17 @@
-import AuthFooter from '../components/auth/AuthFooter';
-import AuthForm from '../components/auth/AuthForm';
-import AuthHeader from '../components/auth/AuthHeader';
-import AuthToggle from '../components/auth/AuthToggle';
-import GoogleAuthButton from '../components/auth/GoogleAuthButton';
-import { useAuth } from '../components/auth/useAuth';
+import AuthFooter from "../components/auth/AuthFooter";
+import AuthForm from "../components/auth/AuthForm";
+import AuthHeader from "../components/auth/AuthHeader";
+import AuthLoading from "../components/auth/AuthLoading";
+import AuthToggle from "../components/auth/AuthToggle";
+import GoogleAuthButton from "../components/auth/GoogleAuthButton";
+import { useAuth } from "../components/auth/useAuth";
 
 const Auth = () => {
   const {
     isLogin,
     setIsLogin,
     isLoading,
+    isGoogleLoading,
     currentForm,
     register,
     errors,
@@ -21,8 +23,11 @@ const Auth = () => {
     <div className="w-full min-h-screen bg-white flex items-center justify-center">
       <div className="m-auto w-[90%] lg:w-1/4 h-auto pb-3">
         <div className="flex flex-col gap-5 items-center">
-          <AuthHeader title={currentForm.title} description={currentForm.disc} />
-          
+          <AuthHeader
+            title={currentForm.title}
+            description={currentForm.disc}
+          />
+
           <AuthToggle isLogin={isLogin} onToggle={setIsLogin} />
 
           <AuthForm
@@ -37,6 +42,7 @@ const Auth = () => {
           <GoogleAuthButton isLogin={isLogin} onClick={handleGoogleLogin} />
 
           <AuthFooter isLogin={isLogin} onToggle={setIsLogin} />
+          {isGoogleLoading && <AuthLoading />}
         </div>
       </div>
     </div>